refactor: isolate chatgpt browser oauth
